Moved link resolvers to libs
